Refactor run to separate file
